Addressed feedback, fixed limit logic
