+++ hedera-mirror-rest/__tests__/specs/topicmessages-01-no-args.spec.json
@@ -1,5 +1,5 @@
 {
-  "description": "Get single message by valid consensus timestamp",
+  "description": "Get topic messages with no arguments",
   "setup": {
     "accounts": [],
     "balances": [],
