+++ hedera-mirror-rest/__tests__/specs/topicmessages-05-all-params.spec.json
@@ -38,7 +38,7 @@
       }
     ]
   },
-  "url": "/api/v1/topic/7?seqnum=gt:3&timestamp=lte:1234567890.000000005",
+  "url": "/api/v1/topic/7?sequencenumber=gt:3&timestamp=lte:1234567890.000000005",
   "responseStatus": 200,
   "responseJson": {
     "messages": [
