+++ hedera-mirror-rest/__tests__/specs/topicmessages-09-limit.spec.json
@@ -38,7 +38,7 @@
       }
     ]
   },
-  "url": "/api/v1/topic/7?seqnum=gt:2&timestamp=lte:1234567890.000000006&limit=2",
+  "url": "/api/v1/topic/7?sequencenumber=gt:2&timestamp=lte:1234567890.000000006&limit=2",
   "responseStatus": 200,
   "responseJson": {
     "messages": [
@@ -58,7 +58,7 @@
       }
     ],
     "links": {
-      "next": "/api/v1/topic/7?seqnum=gt:2&timestamp=lte:1234567890.000000006&timestamp=gt:1234567890.000000004&limit=2"
+      "next": "/api/v1/topic/7?sequencenumber=gt:2&timestamp=lte:1234567890.000000006&timestamp=gt:1234567890.000000004&limit=2"
     }
   }
 }
