+++ hedera-mirror-rest/__tests__/specs/topicmessages-10-order.spec.json
@@ -38,7 +38,7 @@
       }
     ]
   },
-  "url": "/api/v1/topic/7?seqnum=gt:2&timestamp=lte:1234567890.000000006&limit=3&order=desc",
+  "url": "/api/v1/topic/7?sequencenumber=gt:2&timestamp=lte:1234567890.000000006&limit=3&order=desc",
   "responseStatus": 200,
   "responseJson": {
     "messages": [
@@ -65,7 +65,7 @@
       }
     ],
     "links": {
-      "next": "/api/v1/topic/7?seqnum=gt:2&limit=3&order=desc&timestamp=lt:1234567890.000000004"
+      "next": "/api/v1/topic/7?sequencenumber=gt:2&limit=3&order=desc&timestamp=lt:1234567890.000000004"
     }
   }
 }
