+++ hedera-mirror-rest/__tests__/utils.test.js
@@ -22,6 +22,7 @@
 const request = require('supertest');
 const utils = require('../utils.js');
 const constants = require('../constants.js');
+const config = require('../config.js');
 
 describe('Utils getNullableNumber tests', () => {
   test('Verify getNullableNumber returns correct result for 0', () => {
@@ -139,7 +140,7 @@ describe('Utils makeValidationResponse tests', () => {
     expect(val).toStrictEqual({
       isValid: false,
       code: utils.httpStatusCodes.BAD_REQUEST,
-      contents: {_status: {messages: message}}
+      contents: {_status: {messages: message}},
     });
   });
 });
@@ -195,240 +196,62 @@ describe('Utils isValidEntityNum tests', () => {
   });
 });
 
-describe('utils buildComparatorFilter tests', () => {
-  test('Verify buildComparatorFilter for seqnum=lt:2', () => {
-    verifyBuildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, 'lt:2', {
-      key: constants.filterKeys.SEQUENCE_NUMBER,
-      operator: 'lt',
-      value: '2'
-    });
+describe('Utils isValidLimitNum tests', () => {
+  test('Verify invalid for null', () => {
+    expect(utils.isValidLimitNum(null)).toBe(false);
   });
-
-  test('Verify buildComparatorFilter for seqnum=lte:3', () => {
-    verifyBuildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, 'lte:3', {
-      key: constants.filterKeys.SEQUENCE_NUMBER,
-      operator: 'lte',
-      value: '3'
-    });
+  test('Verify invalid for empty input', () => {
+    expect(utils.isValidLimitNum('')).toBe(false);
   });
-
-  test('Verify buildComparatorFilter for seqnum=4', () => {
-    verifyBuildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, '4', {
-      key: constants.filterKeys.SEQUENCE_NUMBER,
-      operator: 'eq',
-      value: '4'
-    });
+  test('Verify invalid for invalid input', () => {
+    expect(utils.isValidLimitNum('1234567890.000000001')).toBe(false);
   });
-
-  test('Verify buildComparatorFilter for timestamp=gte:1234567890.000000004', () => {
-    verifyBuildComparatorFilter(constants.filterKeys.TIMESTAMP, 'gte:1234567890.000000005', {
-      key: constants.filterKeys.TIMESTAMP,
-      operator: 'gte',
-      value: '1234567890.000000005'
-    });
+  test('Verify invalid for entity format shard', () => {
+    expect(utils.isValidLimitNum('1.0.1')).toBe(false);
   });
-
-  test('Verify buildComparatorFilter for timestamp=gt:6', () => {
-    verifyBuildComparatorFilter(constants.filterKeys.TIMESTAMP, 'gt:1234567890.000000006', {
-      key: constants.filterKeys.TIMESTAMP,
-      operator: 'gt',
-      value: '1234567890.000000006'
-    });
+  test('Verify invalid for negative num', () => {
+    expect(utils.isValidLimitNum('-1')).toBe(false);
+  });
+  test('Verify invalid above max limit', () => {
+    expect(utils.isValidLimitNum(config.api.maxLimit + 1)).toBe(false);
+  });
+  test('Verify invalid for 0', () => {
+    expect(utils.isValidLimitNum(0)).toBe(false);
+  });
+  test('Verify valid for valid number', () => {
+    expect(utils.isValidLimitNum(123)).toBe(true);
+  });
+  test(`Verify valid for max limit or ${config.api.maxLimit}`, () => {
+    expect(utils.isValidLimitNum(config.api.maxLimit)).toBe(true);
   });
 });
 
-const verifyBuildComparatorFilter = (key, val, expectedFilter) => {
-  let filter = utils.buildComparatorFilter(key, val);
-
-  verifyFilter(filter, expectedFilter.key, expectedFilter.operator, expectedFilter.value);
-};
-
-const verifyFilter = (filter, key, op, val) => {
-  expect(filter.key).toStrictEqual(key);
-  expect(filter.operator).toStrictEqual(op);
-  expect(filter.value).toStrictEqual(val);
-};
-
-describe('utils buildFilterObject tests', () => {
-  test('Verify buildComparatorFilter for /api/v1/topic/7?seqnum=2', () => {
-    const filters = {
-      seqnum: '2'
-    };
-
-    let formattedFilters = utils.buildFilterObject(filters);
-
-    expect(formattedFilters.length).toBe(1);
-    verifyFilter(formattedFilters[0], constants.filterKeys.SEQUENCE_NUMBER, 'eq', '2');
+describe('Utils isValidNum tests', () => {
+  test('Verify invalid for null', () => {
+    expect(utils.isValidNum(null)).toBe(false);
   });
-
-  test('Verify buildComparatorFilter for /api/v1/topic/7?timestamp=1234567890.000000004', () => {
-    const filters = {
-      timestamp: '1234567890.000000004'
-    };
-
-    let formattedFilters = utils.buildFilterObject(filters);
-
-    expect(formattedFilters.length).toBe(1);
-    verifyFilter(formattedFilters[0], constants.filterKeys.TIMESTAMP, 'eq', '1234567890.000000004');
+  test('Verify invalid for empty input', () => {
+    expect(utils.isValidNum('')).toBe(false);
   });
-
-  test('Verify buildComparatorFilter for /api/v1/topic/7?seqnum=2&seqnum=gte:3&timestamp=1234567890.000000004&order=desc&limit=5', () => {
-    const filters = {
-      seqnum: ['lt:2', 'gte:3'],
-      timestamp: '1234567890.000000004',
-      limit: '5',
-      order: 'desc'
-    };
-
-    let builtFilters = utils.buildFilterObject(filters);
-
-    expect(builtFilters.length).toBe(5);
-    verifyFilter(builtFilters[0], constants.filterKeys.SEQUENCE_NUMBER, 'lt', '2');
-    verifyFilter(builtFilters[1], constants.filterKeys.SEQUENCE_NUMBER, 'gte', '3');
-    verifyFilter(builtFilters[2], constants.filterKeys.TIMESTAMP, 'eq', '1234567890.000000004');
-    verifyFilter(builtFilters[3], constants.filterKeys.LIMIT, 'eq', '5');
-    verifyFilter(builtFilters[4], constants.filterKeys.ORDER, 'eq', 'desc');
+  test('Verify invalid for invalid input', () => {
+    expect(utils.isValidNum('1234567890.000000001')).toBe(false);
+  });
+  test('Verify invalid for entity format shard', () => {
+    expect(utils.isValidNum('1.0.1')).toBe(false);
+  });
+  test('Verify invalid for negative num', () => {
+    expect(utils.isValidNum(-1)).toBe(false);
+  });
+  test('Verify invalid for 0', () => {
+    expect(utils.isValidNum(0)).toBe(false);
+  });
+  test('Verify valid for valid number', () => {
+    expect(utils.isValidNum(123)).toBe(true);
+  });
+  test('Verify valid above max limit', () => {
+    expect(utils.isValidNum(12345678901)).toBe(true);
+  });
+  test(`Verify valid for Number.MAX_SAFE_INTEGER: ${Number.MAX_SAFE_INTEGER}`, () => {
+    expect(utils.isValidNum(Number.MAX_SAFE_INTEGER)).toBe(true);
   });
 });
-
-describe('utils formatComparator tests', () => {
-  test('Verify formatComparator for seqnum=2', () => {
-    const filter = utils.buildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, 'lt:2');
-
-    let formattedFilter = utils.formatComparator(filter);
-
-    expect(formattedFilter).toBe(undefined);
-  });
-
-  test('Verify formatComparator for timestamp=lte:1234567890.000000004', () => {
-    const filter = utils.buildComparatorFilter(constants.filterKeys.TIMESTAMP, 'lte:1234567890.000000004');
-    utils.formatComparator(filter);
-    verifyFilter(filter, constants.filterKeys.TIMESTAMP, ' <= ', '1234567890000000004');
-  });
-
-  test('Verify formatComparator for account.id=5', () => {
-    const filter = utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_ID, '5');
-    utils.formatComparator(filter);
-    verifyFilter(filter, constants.filterKeys.ACCOUNT_ID, ' = ', {num: '5', realm: 0, shard: 0});
-  });
-
-  test('Verify formatComparator for timestamp=1234567890.000000004', () => {
-    const filter = utils.buildComparatorFilter(constants.filterKeys.TIMESTAMP, '1234567890.000000005');
-    utils.formatComparator(filter);
-    verifyFilter(filter, constants.filterKeys.TIMESTAMP, ' = ', '1234567890000000005');
-  });
-
-  test('Verify formatComparator for account.id=gte:6', () => {
-    const filter = utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_ID, 'gte:6');
-    utils.formatComparator(filter);
-    verifyFilter(filter, constants.filterKeys.ACCOUNT_ID, ' >= ', {num: '6', realm: 0, shard: 0});
-  });
-
-  test('Verify formatComparator for timestamp=lte:1234567890.000000004', () => {
-    const filter = utils.buildComparatorFilter(constants.filterKeys.TIMESTAMP, 'gt:1234567890.000000007');
-    utils.formatComparator(filter);
-    verifyFilter(filter, constants.filterKeys.TIMESTAMP, ' > ', '1234567890000000007');
-  });
-});
-
-describe('utils validateAndParseFilters tests', () => {
-  test('Verify validateAndParseFilters for seqnum=2', () => {
-    const filters = [
-      utils.buildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, '<:2'),
-      utils.buildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, '<=:2')
-    ];
-
-    let validationResponse = utils.validateAndParseFilters(filters);
-
-    verifyInvalidFilters(validationResponse, [
-      constants.filterKeys.SEQUENCE_NUMBER,
-      constants.filterKeys.SEQUENCE_NUMBER
-    ]);
-  });
-
-  test('Verify validateAndParseFilters for erroneous data', () => {
-    const filters = [
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_ID, 'lt:-1'),
-      utils.buildComparatorFilter(constants.filterKeys.TIMESTAMP, 'lte:today'),
-      utils.buildComparatorFilter(constants.filterKeys.ORDER, 'chronological'),
-      utils.buildComparatorFilter(constants.filterKeys.LIMIT, '-1'),
-      utils.buildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, '-1'),
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_PUBLICKEY, 'key'),
-      utils.buildComparatorFilter(constants.filterKeys.RESULT, 'good'),
-      utils.buildComparatorFilter(constants.filterKeys.TYPE, 'all'),
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_BALANCE, '-1')
-    ];
-
-    let validationResponse = utils.validateAndParseFilters(filters);
-
-    verifyInvalidFilters(validationResponse, [
-      constants.filterKeys.ACCOUNT_ID,
-      constants.filterKeys.TIMESTAMP,
-      constants.filterKeys.ORDER,
-      constants.filterKeys.LIMIT,
-      constants.filterKeys.SEQUENCE_NUMBER,
-      constants.filterKeys.ACCOUNT_PUBLICKEY,
-      constants.filterKeys.RESULT,
-      constants.filterKeys.TYPE,
-      constants.filterKeys.ACCOUNT_BALANCE
-    ]);
-  });
-
-  test('Verify validateAndParseFilters for invalid format', () => {
-    const filters = [
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_ID, 'lt:0.1.23456789012345'),
-      utils.buildComparatorFilter(constants.filterKeys.TIMESTAMP, 'lte:23456789012345678901234'),
-      utils.buildComparatorFilter(constants.filterKeys.LIMIT, '2.3'),
-      utils.buildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, '3.4'),
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_PUBLICKEY, '3c3d546321ff6f63d701d2ec5c2'),
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_BALANCE, '23456789012345678901234')
-    ];
-
-    let validationResponse = utils.validateAndParseFilters(filters);
-
-    verifyInvalidFilters(validationResponse, [
-      constants.filterKeys.ACCOUNT_ID,
-      constants.filterKeys.TIMESTAMP,
-      constants.filterKeys.LIMIT,
-      constants.filterKeys.SEQUENCE_NUMBER,
-      constants.filterKeys.ACCOUNT_PUBLICKEY,
-      constants.filterKeys.ACCOUNT_BALANCE
-    ]);
-  });
-
-  test('Verify validateAndParseFilters for valid filters', () => {
-    const filters = [
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_ID, 'lt:2'),
-      utils.buildComparatorFilter(constants.filterKeys.TIMESTAMP, 'lte:1234567890.000000003'),
-      utils.buildComparatorFilter(constants.filterKeys.ORDER, 'desc'),
-      utils.buildComparatorFilter(constants.filterKeys.LIMIT, 'gte:4'),
-      utils.buildComparatorFilter(constants.filterKeys.SEQUENCE_NUMBER, '5'),
-      utils.buildComparatorFilter(
-        constants.filterKeys.ACCOUNT_PUBLICKEY,
-        '3c3d546321ff6f63d701d2ec5c277095874e19f4a235bee1e6bb19258bf362be'
-      ),
-      utils.buildComparatorFilter(constants.filterKeys.RESULT, 'success'),
-      utils.buildComparatorFilter(constants.filterKeys.RESULT, 'fail'),
-      utils.buildComparatorFilter(constants.filterKeys.TYPE, 'credit'),
-      utils.buildComparatorFilter(constants.filterKeys.TYPE, 'debit'),
-      utils.buildComparatorFilter(constants.filterKeys.ACCOUNT_BALANCE, '45000')
-    ];
-
-    let validationResponse = utils.validateAndParseFilters(filters);
-
-    verifyValidFilters(validationResponse);
-  });
-});
-
-const verifyValidFilters = val => {
-  expect(val).toStrictEqual(utils.successValidationResponse);
-};
-
-const verifyInvalidFilters = (val, columns) => {
-  let invalidObjects = [];
-  for (let col of columns) {
-    invalidObjects.push(utils.getInvalidParameterMessageObject(col));
-  }
-
-  expect(val).toStrictEqual(utils.makeValidationResponse(invalidObjects));
-};
