+++ hedera-mirror-rest/constants.js
@@ -26,19 +26,19 @@ const filterKeys = {
   LIMIT: 'limit',
   ORDER: 'order',
   RESULT: 'result',
-  SEQUENCE_NUMBER: 'seqnum',
+  SEQUENCE_NUMBER: 'sequencenumber',
   TIMESTAMP: 'timestamp',
-  TYPE: 'type'
+  TYPE: 'type',
 };
 
 const entityColumns = {
   ENTITY_NUM: 'entity_num',
   ENTITY_REALM: 'entity_realm',
   ENTITY_SHARD: 'entity_shard',
-  PUBLIC_KEY: 'ed25519_public_key_hex'
+  PUBLIC_KEY: 'ed25519_public_key_hex',
 };
 
 module.exports = {
   entityColumns: entityColumns,
-  filterKeys: filterKeys
+  filterKeys: filterKeys,
 };
