+++ hedera-mirror-rest/server.js
@@ -115,8 +115,8 @@ app.get(apiPrefix + '/topic/message/:consensusTimestamp', topicmessage.getMessag
 app.use(utils.errorHandler);
 
 // support singular and plural resource naming for single topic message via id and sequence
-app.get(apiPrefix + '/topic/:id/message/:seqnum', topicmessage.getMessageByTopicAndSequenceRequest);
-app.get(apiPrefix + '/topics/:id/messages/:seqnum', topicmessage.getMessageByTopicAndSequenceRequest);
+app.get(apiPrefix + '/topic/:id/message/:sequencenumber', topicmessage.getMessageByTopicAndSequenceRequest);
+app.get(apiPrefix + '/topics/:id/messages/:sequencenumber', topicmessage.getMessageByTopicAndSequenceRequest);
 
 app.get(apiPrefix + '/topics/:id', topicmessage.getTopicMessages);
 app.get(apiPrefix + '/topic/:id', topicmessage.getTopicMessages);
